Move switchServer code into configureInterface
