+++ src/platforms/linux/daemon/wgutilslinux.h
@@ -5,6 +5,7 @@
 #ifndef WIREGUARDUTILSLINUX_H
 #define WIREGUARDUTILSLINUX_H
 
+#include "daemon/daemon.h"
 #include "daemon/wgutils.h"
 #include <QStringList>
 
@@ -17,6 +18,15 @@ class WireguardUtilsLinux final : public WireguardUtils {
   QStringList currentInterfaces() override;
   bool removeInterfaceIfExists();
   peerBytes getThroughputForInterface();
+  bool configureInterface(const Daemon::Config& config);
+
+ private:
+  bool setPeerEndpoint(struct sockaddr* peerEndpoint, const QString& address,
+                       int port);
+  bool setAllowedIpsOnPeer(struct wg_peer* peer,
+                           QList<IPAddressRange> allowedIPAddressRanges);
+  wg_peer* buildPeerForDevice(struct wg_device* device,
+                              const Daemon::Config& conf);
 };
 
 #endif  // WIREGUARDUTILSLINUX_H
